+++ test/src/updater/blackboxUpdateWinSuite.ts
@@ -1,11 +1,14 @@
-import { Arch } from "electron-builder"
 import { ToolsetConfig } from "app-builder-lib"
+import { Arch } from "electron-builder"
 import { TestContext } from "vitest"
-import { runTest } from "./blackboxUpdateHelpers"
+import { optionsForFlakyE2E, runTest, windowsVmPromise } from "./blackboxUpdateHelpers"
 
 export function registerBlackboxWinTests(winCodeSign: ToolsetConfig["winCodeSign"]): void {
-  describe.heavy.ifWindows("windows", { sequential: true, retry: 2 }, () => {
+  describe.heavy("windows", optionsForFlakyE2E, () => {
     test("nsis", async (context: TestContext) => {
+      if (process.platform !== "win32" && (await windowsVmPromise) == null) {
+        context.skip()
+      }
       await runTest(context, "nsis", "", Arch.x64, { winCodeSign })
     })
   })
